Rename Unit to Unit Service, and Unit Service to Unit Manager Service
