--- Assets/Scripts/GamePlay/Services/Interfaces/IUnitService.cs
@@ -1,13 +1,0 @@
-using System;
-using UnityEngine;
-using System.Collections.Generic;
-
-namespace Source.GamePlay.Services.Interfaces
-{
-    public interface IUnitService
-    {
-        void SelectUnit(Unit selectedUnit, List<Unit> units);
-        void SelectUnits(Guid playerId, Vector3 selectionStart, Vector3 selectionEnd, List<Unit> units);
-        void MoveUnits(Guid playerId, Vector3 destination, List<Unit> units);
-    }
-}
