--- Assets/Scripts/GamePlay/Services/Unit.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5f3f0f54456e7094e8ca2cf122515f5c
